requirements.txt
fastapi
uvicorn
databases
asyncpg
python-dotenv
psycopg2-binary

